renamed entry db to entrydatabase
