+++ src/DataBase/EntryDatabase.cs
@@ -1,21 +1,25 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
 using ePiggyWeb.DataBase.Models;
 using ePiggyWeb.DataManagement;
 using ePiggyWeb.DataManagement.Entries;
 using ePiggyWeb.Utilities;
+using Microsoft.EntityFrameworkCore;
 
 namespace ePiggyWeb.DataBase
 {
-    public static class EntryDatabase
+    public class EntryDatabase
     {
-        /*
-        I've made all methods here require an EntryType instead of having to call different Add Income or Add Expense methods everywhere else
-        Also moved to a standard of just passing IEntry everywhere, instead of many types of different parameters, because that quickly clogged up the code in this class.
-         */
+        private PiggyDbContext Database { get; }
+        public EntryDatabase(PiggyDbContext database)
+        {
+            Database = database;
+        }
 
-        public static int Create(IEntry localEntry, int userId, EntryType entryType)
+        public async Task<int> CreateAsync(IEntry localEntry, int userId, EntryType entryType)
         {
             if (localEntry is null)
             {
@@ -24,29 +28,27 @@ namespace ePiggyWeb.DataBase
 
             if (!localEntry.Recurring)
             {
-                return CreateSingle(localEntry, userId, entryType);
+                return await CreateSingleAsync(localEntry, userId, entryType);
             }
 
-            CreateList(RecurringUpdater.CreateRecurringList(localEntry, entryType), userId);
+            await CreateListAsync(RecurringUpdater.CreateRecurringList(localEntry, entryType), userId);
             return -2;
         }
 
-        public static int CreateSingle(IEntry localEntry, int userId, EntryType entryType)
+        public async Task<int> CreateSingleAsync(IEntry localEntry, int userId, EntryType entryType)
         {
-            using var db = new DatabaseContext();
             IEntryModel dbEntry = entryType switch
             {
                 EntryType.Income => new IncomeModel(localEntry, userId),
                 _ => new ExpenseModel(localEntry, userId),
             };
-            db.Add(dbEntry);
-            db.SaveChanges();
+            await Database.AddAsync(dbEntry);
+            await Database.SaveChangesAsync();
             return dbEntry.Id;
         }
 
-        public static bool CreateList(IEntryList entryList, int userId)
+        public async Task<bool> CreateListAsync(IEntryList entryList, int userId)
         {
-            using var db = new DatabaseContext();
             var dictionary = new Dictionary<IEntry, IEntryModel>();
             foreach (var entry in entryList)
             {
@@ -58,8 +60,8 @@ namespace ePiggyWeb.DataBase
                 dictionary.Add(entry, dbEntry);
             }
             // Setting all of the ID's to local Entries, just so this method remains usable both with local and only database usage
-            db.AddRange(dictionary.Values);
-            db.SaveChanges();
+            await Database.AddRangeAsync(dictionary.Values);
+            await Database.SaveChangesAsync();
             entryList.Clear();
             foreach (var (key, value) in dictionary)
             {
@@ -70,7 +72,8 @@ namespace ePiggyWeb.DataBase
             return true;
         }
 
-        public static bool Update(int id, int userId, IEntry updatedEntry, EntryType entryType)
+
+        public async Task<bool> UpdateAsync(int id, int userId, IEntry updatedEntry, EntryType entryType)
         {
             if (updatedEntry is null)
             {
@@ -79,20 +82,19 @@ namespace ePiggyWeb.DataBase
 
             if (updatedEntry.Recurring)
             {
-                CreateList(RecurringUpdater.CreateRecurringListWithoutOriginalEntry(updatedEntry, entryType), userId);
+                await CreateListAsync(RecurringUpdater.CreateRecurringListWithoutOriginalEntry(updatedEntry, entryType), userId);
                 updatedEntry.Recurring = false;
             }
 
-            return UpdateSingle(id, userId, updatedEntry, entryType);
+            return await UpdateSingleAsync(id, userId, updatedEntry, entryType);
         }
 
-        public static bool UpdateSingle(int id, int userId, IEntry updatedEntry, EntryType entryType)
+        public async Task<bool> UpdateSingleAsync(int id, int userId, IEntry updatedEntry, EntryType entryType)
         {
-            using var db = new DatabaseContext();
             IEntryModel temp = entryType switch
             {
-                EntryType.Income => db.Incomes.FirstOrDefault(x => x.Id == id && x.UserId == userId),
-                _ => db.Expenses.FirstOrDefault(x => x.Id == id && x.UserId == userId)
+                EntryType.Income => await Database.Incomes.FirstOrDefaultAsync(x => x.Id == id && x.UserId == userId),
+                _ => await Database.Expenses.FirstOrDefaultAsync(x => x.Id == id && x.UserId == userId)
             };
 
             if (temp is null)
@@ -103,32 +105,31 @@ namespace ePiggyWeb.DataBase
 
             updatedEntry.UserId = userId;
             temp.Edit(updatedEntry);
-            db.SaveChanges();
+            await Database.SaveChangesAsync();
             return true;
         }
-
-        public static bool Delete(IEntry entry, EntryType entryType)
+        
+        public async Task<bool> DeleteAsync(IEntry entry, EntryType entryType)
         {
-            return Delete(x => x.Id == entry.Id && x.UserId == entry.UserId, entryType);
+            return await DeleteAsync(x => x.Id == entry.Id && x.UserId == entry.UserId, entryType);
         }
 
-        public static bool Delete(int id, int userId, EntryType entryType)
+        public async Task<bool> DeleteAsync(int id, int userId, EntryType entryType)
         {
-            return Delete(x => x.Id == id && x.UserId == userId, entryType);
+            return await DeleteAsync(x => x.Id == id && x.UserId == userId, entryType);
         }
 
-        public static bool Delete(Func<IEntryModel, bool> filter, EntryType entryType)
+        public async Task<bool> DeleteAsync(Expression<Func<IEntryModel, bool>> filter, EntryType entryType)
         {
-            using var db = new DatabaseContext();
             try
             {
                 var dbEntry = entryType switch
                 {
-                    EntryType.Income => db.Incomes.FirstOrDefault(filter),
-                    _ => db.Expenses.FirstOrDefault(filter)
+                    EntryType.Income => Database.Incomes.FirstOrDefault(filter),
+                    _ => Database.Expenses.FirstOrDefault(filter)
                 };
-                db.Remove(dbEntry ?? throw new InvalidOperationException());
-                db.SaveChanges();
+                Database.Remove(dbEntry ?? throw new InvalidOperationException());
+                await Database.SaveChangesAsync();
             }
             catch (InvalidOperationException ex)
             {
@@ -139,69 +140,86 @@ namespace ePiggyWeb.DataBase
             return true;
         }
 
-        public static bool DeleteList(Func<IEntryModel, bool> filter, EntryType entryType)
+        public async Task<bool> DeleteListAsync(Expression<Func<IEntryModel, bool>> filter, EntryType entryType)
         {
-            var db = new DatabaseContext();
-
             if (entryType == EntryType.Income)
             {
-                var entriesToRemove = db.Incomes.Where(filter).Cast<IncomeModel>().ToList();
-                db.Incomes.RemoveRange(entriesToRemove);
+                var entriesToRemove = await Database.Incomes.Where(filter).Cast<IncomeModel>().ToListAsync();
+                Database.Incomes.RemoveRange(entriesToRemove);
             }
             else
             {
-                var entriesToRemove = db.Expenses.Where(filter).Cast<ExpenseModel>().ToList();
-                db.Expenses.RemoveRange(entriesToRemove);
+                var entriesToRemove = await Database.Expenses.Where(filter).Cast<ExpenseModel>().ToListAsync();
+                Database.Expenses.RemoveRange(entriesToRemove);
             }
 
-            db.SaveChanges();
+            await Database.SaveChangesAsync();
             return true;
         }
 
-        public static bool DeleteList(IEnumerable<int> idArray, int userId, EntryType entryType)
+        public async Task<bool> DeleteListAsync(IEnumerable<int> idArray, int userId, EntryType entryType)
         {
-            var db = new DatabaseContext();
-
             if (entryType == EntryType.Income)
             {
-                var entriesToRemove = idArray.Select(id => db.Incomes.FirstOrDefault(x => x.Id == id && x.UserId == userId)).ToList();
-                db.Incomes.RemoveRange(entriesToRemove);
+                var list = new List<IncomeModel>();
+                foreach (var id in idArray)
+                {
+                    var temp = await Database.Incomes.FirstOrDefaultAsync(x => x.Id == id && x.UserId == userId);
+                    list.Add(temp);
+                }
+                Database.Incomes.RemoveRange(list);
             }
             else
             {
-                var entriesToRemove = idArray.Select(id => db.Expenses.FirstOrDefault(x => x.Id == id && x.UserId == userId)).ToList();
-                db.Expenses.RemoveRange(entriesToRemove);
+                var list = new List<ExpenseModel>();
+                foreach (var id in idArray)
+                {
+                    var temp = await Database.Expenses.FirstOrDefaultAsync(x => x.Id == id && x.UserId == userId);
+                    list.Add(temp);
+                }
+                Database.Expenses.RemoveRange(list);
             }
 
-            db.SaveChanges();
+            await Database.SaveChangesAsync();
             return true;
         }
 
-        public static IEntry Read(int id, int userId, EntryType entryType)
+
+        public async Task<IEntry> ReadAsync(int id, int userId, EntryType entryType)
         {
-            using var db = new DatabaseContext();
             IEntryModel dbEntry = entryType switch
             {
-                EntryType.Income => db.Incomes.FirstOrDefault(x => x.Id == id && x.UserId == userId),
-                _ => db.Expenses.FirstOrDefault(x => x.Id == id && x.UserId == userId)
+                EntryType.Income => Database.Incomes.FirstOrDefault(x => x.Id == id && x.UserId == userId),
+                _ => Database.Expenses.FirstOrDefault(x => x.Id == id && x.UserId == userId)
+            };
+            return await ReadAsync(x => x.Id == id && x.UserId == userId, entryType);
+        }
+
+        public async Task<IEntry> ReadAsync(Expression<Func<IEntryModel, bool>> filter, EntryType entryType)
+        {
+            var dbEntry = entryType switch
+            {
+                EntryType.Income => await Database.Incomes.FirstOrDefaultAsync(filter),
+                _ => await Database.Expenses.FirstOrDefaultAsync(filter)
             };
             return new Entry(dbEntry);
         }
 
-        public static IEnumerable<IEntry> ReadList(int userId, EntryType entryType)
+        public async Task<IEntryList> ReadListAsync(int userId, EntryType entryType)
         {
-            return ReadList(x => x.UserId == userId, entryType);
+            return await ReadListAsync(x => x.UserId == userId, entryType);
         }
 
-        public static IEnumerable<IEntry> ReadList(Func<IEntryModel, bool> filter, EntryType entryType)
+        public async Task<IEntryList> ReadListAsync(Expression<Func<IEntryModel, bool>> filter, EntryType entryType)
         {
-            using var db = new DatabaseContext();
+            var list = new EntryList(entryType);
             IEnumerable<IEntry> temp = entryType switch
             {
-                EntryType.Income => db.Incomes.Where(filter).Select(dbEntry => new Entry(dbEntry)).Cast<IEntry>().ToList(),
-                _ => db.Expenses.Where(filter).Select(dbEntry => new Entry(dbEntry)).Cast<IEntry>().ToList()
+                EntryType.Income => await Database.Incomes.Where(filter).Select(dbEntry => new Entry(dbEntry)).Cast<IEntry>().ToListAsync(),
+                _ => await Database.Expenses.Where(filter).Select(dbEntry => new Entry(dbEntry)).Cast<IEntry>().ToListAsync()
             };
-            return temp;
+            list.AddRange(temp);
+            return list;
         }
     }
 }
